Switched the light-cone array to be a 32bit float (equiavalent to the C-data used to populate it)

+++ src/py21cmmc/_21cmfast/wrapper.py
@@ -1734,7 +1734,7 @@ def run_lightcone(*, redshift=None, max_redshift=None, user_params=None, cosmo_p
     d_at_redshift, lc_distances, n_lightcone = _setup_lightcone(cosmo_params, max_redshift, redshift, scrollz,
                                                                 user_params, global_params.ZPRIME_STEP_FACTOR)
 
-    lc = np.zeros((user_params.HII_DIM, user_params.HII_DIM, n_lightcone))
+    lc = np.zeros((user_params.HII_DIM, user_params.HII_DIM, n_lightcone), dtype=np.float32)
 
     scroll_distances = cosmo_params.cosmo.comoving_distance(scrollz).value - d_at_redshift
